Filter GTFS data by trip_id

Filters a GTFS object by trip_ids, keeping (or dropping) the entries that relate to such ids in each file.

filter_trip_id(gtfs, trip_id, keep = TRUE)

Arguments

gtfs
A GTFS object.

trip_id
A character vector. The trip_ids used to filter the data.

keep
A logical. Whether the entries related to the passed trip_ids should be kept or dropped (defaults to TRUE, which keeps the entries).

Value

The GTFS object passed to the gtfs parameter, after the filtering process.

Examples

data_path <- system.file("extdata/spo_gtfs.zip", package = "gtfstools")
gtfs <- read_gtfs(data_path)
trip_ids <- c("CPTM L07-0", "2002-10-0")

object.size(gtfs)

#> 811232 bytes

# keeps entries related to passed trip_ids
smaller_gtfs <- filter_trip_id(gtfs, trip_ids)
object.size(smaller_gtfs)

#> 72984 bytes

# drops entries related to passed trip_ids
smaller_gtfs <- filter_trip_id(gtfs, trip_ids, keep = FALSE)
object.size(smaller_gtfs)

#> 768720 bytes