gtfstools: General Transit Feed Specification (GTFS) Editing and Analysing Tools

Utility functions to read, manipulate, analyse and write transit feeds in the General Transit Feed Specification (GTFS) data format.

Usage

Please check the vignettes for more on the package usage:

Basic usage: reading, analysing, manipulating and writing feeds. Run vignette("gtfstools") or check it on the website.

Filtering GTFS feeds. Run vignette("filtering", package = "gtfstools") or check it on the website.

Validating GTFS feeds. Run vignette("validating", package = "gtfstools") or check it on the website.

See also

Author

Maintainer: Daniel Herszenhut dhersz@gmail.com (ORCID)

Authors:

Rafael H. M. Pereira (ORCID)

Pedro R. Andrade (ORCID)

Joao Bazzo (ORCID)

Other contributors:

Mark Padgham [contributor]

Marcus Saraiva (ORCID) [contributor]

Ipea - Institute for Applied Economic Research [copyright holder, funder]